Extract api integration to a external service
